AB#1011 Move page to proper directory
