+++ client/tests/webdriver/pages/showMartImportedReports.page.js
@@ -1,4 +1,4 @@
-import Page from "../page"
+import Page from "./page"
 
 const PAGE_URL = "/admin/martImportedReports"
